requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
autopep8==1.5.3
bandit==1.6.2
billiard==3.6.3.0
black==19.10b0
certifi==2020.6.20
cffi==1.14.0
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
distlib==0.3.1
Django==3.1
django-debug-toolbar==2.2
django-filter==2.3.0
djangorestframework==3.11.1
docutils==0.15.2
dparse==0.5.1
drf-yasg==1.17.1
filelock==3.0.12
flake8
gitdb==4.0.5
GitPython==3.1.7
gunicorn==20.0.4
identify==1.4.28
idna==2.10
importlib-metadata
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
Markdown
MarkupSafe==1.1.1
mccabe==0.6.1
nodeenv==1.4.0
packaging==20.4
pathspec==0.8.0
pbr==5.4.5
pre-commit
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pycurl==7.43.0.5
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
PyYAML
regex==2020.7.14
requests==2.24.0
safety==1.9.0
simplejson==3.17.2
six==1.15.0
smmap==3.0.4
sqlparse==0.3.1
stevedore
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.31
zipp==3.1.0
user-agents==2.2.0
django-cors-headers==3.7.0